SCRUM-172 refactor + added timer handling
