+++ app/services/timer.py
@@ -5,23 +5,26 @@ from app.services import game
 
 TURN_TIME_LIMIT = 120
 
-# Global dictionary to keep track of timer tasks
-timer_tasks = {}
+timer_tasks = {}    # Global dictionary to keep track of timer tasks
 
 
-async def handle_timer(game_id, player_id, db):
+def stop_timer(game_id):
 
-    # Cancel the previous timer task if it exists
     if game_id in timer_tasks:
-        # Cancel the previous timer task
         timer_tasks[game_id].cancel()
         del timer_tasks[game_id]
 
-    # Start the timer for the new turn
+
+def start_timer(game_id, player_id, db):
+
     timer_tasks[game_id] = asyncio.create_task(
         emit_timer(game_id, player_id, db)
     )
 
+async def handle_timer(game_id, player_id, db):
+
+    stop_timer(game_id)
+    start_timer(game_id, player_id, db)
 
 async def emit_timer(game_id, player_id, db):
